requirements.txt
# ===============================
# Core API / Server
# ===============================
fastapi>=0.128,<1.0
uvicorn[standard]>=0.40,<1.0
httpx>=0.28,<1.0
python-multipart>=0.0.22,<1.0

# ===============================
# Config / Env
# ===============================
pydantic>=2.12,<3.0
pydantic-settings>=2.12,<3.0
python-dotenv>=1.2,<2.0

# ===============================
# Audio Processing (STT/TTS common)
# ===============================
numpy>=1.26,<2.0
soundfile>=0.13,<1.0
librosa>=0.11,<1.0

# ===============================
# Google Cloud (STT / TTS)
# ===============================
google-cloud-speech>=2.36,<3.0
google-cloud-texttospeech>=2.34,<3.0
google-auth>=2.47,<3.0

# ===============================
# LangGraph / RAG
# ===============================
langgraph>=1.0,<2.0
langchain>=1.2,<2.0
langchain-community>=0.4,<1.0
sentence-transformers>=2.7,<3.0
pypdf>=4.3,<5.0
pdfplumber>=0.11,<1.0
openai>=2.17,<3.0

# ===============================
# Vector DB / Postgres
# ===============================
psycopg[binary]>=3.3,<4.0
pgvector>=0.4,<1.0

# ===============================
# Optional (로컬 FAISS 쓸 때만)
# ===============================
faiss-cpu>=1.13,<2.0
